refactor(api): move festivals to src/api with queryOptions factories
Mechanical move of the festivals feature from src/hooks/queries/festivals
to src/api/festivals; adds per-file queryOptions factories for the query
hooks and repoints all importers. No behavior change. Part of #52.
