+++ src/hooks/queries/festivals/editions/useDeleteFestivalEdition.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { festivalsKeys } from "../types";
+import { festivalsKeys } from "@/api/festivals/types";
 
 async function deleteFestivalEdition(editionId: string) {
   const { error } = await supabase
