+++ src/hooks/queries/festivals/editions/useUpdateFestivalEdition.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/integrations/supabase/client";
-import { festivalsKeys } from "../types";
+import { festivalsKeys } from "@/api/festivals/types";
 import type { Database } from "@/integrations/supabase/types";
 
 type UpdateEditionData =
