+++ src/pages/admin/festivals/FestivalDialog.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
-import { useCreateFestivalMutation } from "@/hooks/queries/festivals/useCreateFestival";
-import { useUpdateFestivalMutation } from "@/hooks/queries/festivals/useUpdateFestival";
-import { Festival } from "@/hooks/queries/festivals/types";
+import { useCreateFestivalMutation } from "@/api/festivals/useCreateFestival";
+import { useUpdateFestivalMutation } from "@/api/festivals/useUpdateFestival";
+import { Festival } from "@/api/festivals/types";
 import { useToast } from "@/hooks/use-toast";
 import {
   Dialog,
